+++ packages/xchain-thorchain/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### Update
 
-- Temporarily hardcodes `fee.gas` in deposit function since the value returned from buildDepositTx is too high, causing transactions from users with sufficient RUNE being rejected.
+- Set `fee.gas` to `auto` in `deposit`
+- Try sending `deposit` tx up to 3x
 - Updates `DEFAULT_GAS_VALUE` to `2000000`
 
 # v.0.13.3 (2021-04-12)